+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Added
 
+- Changed Vizzu's version to 0.7.x.
 - Added: `__all__` into `__init__.py`
 
 ## [0.13.0] - 2022-10-19